Move pkg to internal/domain and adapters to internal/adapters
