+++ internal/adapters/log/group_logger.go
@@ -1,7 +1,7 @@
 package log
 
 import (
-	"thop/pkg/log"
+	"thop/internal/domain/log"
 )
 
 type GroupLogger struct {
