+++ internal/adapters/template/yaml_config.go
