+++ internal/adapters/template/yaml_template.go
