+++ internal/domain/template/storage.go
